+++ packages/dotdart/test/widgets/lottie_generator_widget_test.dart
@@ -81,6 +81,7 @@ void main() {
       },
     );
   });
+
 }
 
 class _MockRectPainter extends CustomPainter {